Minor improvements to make_formatted_string.
* alloc.c (make_formatted_string): Prefer int to ptrdiff_t
where int is good enough, as vsprintf returns an int.
* lisp.h (make_formatted_string): Add ATTRIBUTE_FORMAT_PRINTF.

+++ src/alloc.c
@@ -2524,7 +2524,7 @@ Lisp_Object
 make_formatted_string (char *buf, const char *format, ...)
 {
   va_list ap;
-  ptrdiff_t length;
+  int length;
 
   va_start (ap, format);
   length = vsprintf (buf, format, ap);
